requirements.txt
python-telegram-bot==20.0
wikipedia-api
python-dotenv

